Don't delete table while running tests in gh action

+++ Tests/AWSSDKSwiftTests/Services/DynamoDB/DynamoDBCodableTests.swift
@@ -59,9 +59,13 @@ final class DynamoDBCodableTests: XCTestCase {
         }
     }
     
-    func deleteTable(name: String) -> EventLoopFuture<DynamoDB.DeleteTableOutput> {
+    func deleteTable(name: String) -> EventLoopFuture<Void> {
+        // for some reason delete table is not working while running within a github action. works everywhere else
+        guard ProcessInfo.processInfo.environment["GITHUB_ACTIONS"] != "true"  else {
+             return Self.dynamoDB.client.eventLoopGroup.next().makeSucceededFuture(())
+        }
         let input = DynamoDB.DeleteTableInput(tableName: name)
-        return Self.dynamoDB.deleteTable(input)
+        return Self.dynamoDB.deleteTable(input).map { _ in }
     }
     
     //MARK: Tests
